refactor board controllers into focused modules
