+++ controllers/board_notes_controller.py
@@ -1,0 +1,146 @@
+from __future__ import annotations
+
+from typing import Optional
+
+from PySide6 import QtCore, QtGui, QtWidgets
+
+from core.board_scene.dialogs import NoteTextEditor, PopupOutsideCloseFilter
+from core.board_scene.items import BoardNoteItem
+
+
+class BoardNotesController:
+    """Owns Board note creation and note editing popup UI."""
+
+    def __init__(self, board_controller: object) -> None:
+        self.board = board_controller
+        self.w = board_controller.w
+
+    def add_note(self) -> None:
+        self.add_note_at(None)
+
+    def add_note_at(self, scene_pos: Optional[QtCore.QPointF]) -> None:
+        board = self.board
+        if not board._project_root:
+            board._notify("Select a project first.")
+            return
+        item = BoardNoteItem("New note...")
+        item.setData(0, "note")
+        if scene_pos is None:
+            scene_pos = board._current_view_scene_center()
+        item.setPos(scene_pos)
+        item.setSelected(True)
+        board._scene.addItem(item)
+        board._commit_scene_mutation(history=False, update_groups=True)
+        self.edit_note(item)
+
+    def edit_note(self, item: BoardNoteItem, global_pos: Optional[QtCore.QPoint] = None) -> None:
+        board = self.board
+        dialog = QtWidgets.QDialog(self.w)
+        dialog.setWindowFlags(QtCore.Qt.WindowType.Tool | QtCore.Qt.WindowType.FramelessWindowHint)
+        dialog.setAttribute(QtCore.Qt.WidgetAttribute.WA_DeleteOnClose, True)
+        dialog.setMinimumWidth(320)
+        layout = QtWidgets.QVBoxLayout(dialog)
+        layout.setContentsMargins(10, 10, 10, 10)
+        layout.setSpacing(8)
+
+        text_edit = NoteTextEditor()
+        text_edit.set_text(item.text_item.toPlainText())
+        text_edit.setMinimumHeight(140)
+        layout.addWidget(text_edit, 1)
+
+        options = QtWidgets.QHBoxLayout()
+        layout.addLayout(options)
+
+        size_label = QtWidgets.QLabel("Font")
+        options.addWidget(size_label)
+        size_spin = QtWidgets.QSpinBox()
+        size_spin.setRange(8, 64)
+        size_spin.setValue(item.note_data().get("font_size", 12))
+        options.addWidget(size_spin)
+
+        align_label = QtWidgets.QLabel("Align")
+        options.addWidget(align_label)
+        align_combo = QtWidgets.QComboBox()
+        align_combo.addItems(["Left", "Center"])
+        align_combo.setCurrentText("Center" if item.note_data().get("align") == "center" else "Left")
+        options.addWidget(align_combo)
+
+        color_btn = QtWidgets.QPushButton("Background")
+        options.addWidget(color_btn)
+        color_preview = QtWidgets.QFrame()
+        color_preview.setFixedSize(24, 24)
+        color_preview.setStyleSheet(f"background: {item.note_data().get('bg', '#99000000')}; border: 1px solid #333;")
+        options.addWidget(color_preview)
+        options.addStretch(1)
+
+        selected_color = QtGui.QColor(item.note_data().get("bg", "#99000000"))
+        applied = False
+
+        def pick_color() -> None:
+            nonlocal selected_color
+            popup_filter.block_outside_close = True
+            color = QtWidgets.QColorDialog.getColor(selected_color, self.w, "Pick background color")
+            popup_filter.block_outside_close = False
+            if color.isValid():
+                selected_color = color
+                color_preview.setStyleSheet(
+                    f"background: {color.name(QtGui.QColor.NameFormat.HexArgb)}; border: 1px solid #333;"
+                )
+                preview_align = (
+                    QtCore.Qt.AlignmentFlag.AlignHCenter
+                    if align_combo.currentText() == "Center"
+                    else QtCore.Qt.AlignmentFlag.AlignLeft
+                )
+                item.set_note_style(size_spin.value(), preview_align, selected_color)
+                board._dirty = True
+
+        def apply_changes() -> None:
+            nonlocal applied
+            if applied:
+                return
+            if item.scene() is None:
+                return
+            applied = True
+            align_flag = (
+                QtCore.Qt.AlignmentFlag.AlignHCenter
+                if align_combo.currentText() == "Center"
+                else QtCore.Qt.AlignmentFlag.AlignLeft
+            )
+            item.set_text(text_edit.text())
+            item.set_note_style(size_spin.value(), align_flag, selected_color)
+            board._commit_scene_mutation(history=True, update_groups=True)
+
+        color_btn.clicked.connect(pick_color)
+        dialog.finished.connect(lambda _result: apply_changes())
+
+        popup_filter = PopupOutsideCloseFilter(dialog, apply_changes)
+        QtWidgets.QApplication.instance().installEventFilter(popup_filter)
+
+        def cleanup() -> None:
+            app = QtWidgets.QApplication.instance()
+            if app is not None:
+                app.removeEventFilter(popup_filter)
+
+        dialog.finished.connect(lambda _result: cleanup())
+
+        if global_pos is None:
+            view = self.w.board_page.view
+            scene_pos = item.sceneBoundingRect().topLeft()
+            view_pos = view.mapFromScene(scene_pos)
+            global_pos = view.viewport().mapToGlobal(view_pos)
+        dialog.adjustSize()
+        target = global_pos + QtCore.QPoint(12, 12)
+        screen = QtGui.QGuiApplication.screenAt(target)
+        if screen is None:
+            screen = QtGui.QGuiApplication.primaryScreen()
+        if screen is not None:
+            geom = screen.availableGeometry()
+            dlg = dialog.frameGeometry()
+            x = max(geom.left() + 8, min(target.x(), geom.right() - dlg.width() - 8))
+            y = max(geom.top() + 8, min(target.y(), geom.bottom() - dlg.height() - 8))
+            target = QtCore.QPoint(x, y)
+        dialog.move(target)
+        dialog.show()
+        dialog.raise_()
+        dialog.activateWindow()
+        QtCore.QTimer.singleShot(0, text_edit.setFocus)
